+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 2.0.8 (in development)
 
-- Swap accidental and octave characters in displays that can print notes (FourView, Part, CVPad, WriteSeq32/64, etc.)
+- Swap accidental and octave characters in displays that can print notes (FourView, Part, CVPad, WriteSeq64, Foundry, but not WriteSeq32 to keep its display more readable)
 
 
 ### 2.0.7 (2021-12-26)